+++ src/main/java/org/truffleruby/core/regexp/RegexWarnDeferredCallback.java
@@ -9,7 +9,7 @@
  */
 package org.truffleruby.core.regexp;
 
-import org.graalvm.shadowed.org.joni.WarnCallback;
+import dev.truffleruby.shadowed.org.joni.WarnCallback;
 import org.truffleruby.parser.RubyDeferredWarnings;
 
 public final class RegexWarnDeferredCallback implements WarnCallback {